Support simple C++ linters
Support basic C++ linters/formatters, such as clang-format or cpplint.
Tools that require a compile-commands.json are not yet in scope.

Clang-format will insist that the configuration file is called
".clang-format" and that it resides in a parent directory, and will
not tolerate any mention of the file on the command line. It wants the
"-i" flag to format code in place. To accomodate this, permit
specifying both config_str and config, and don't mention the config
file on the command line unless config_opt is specified. In the
clang-format case, you would run with config_str="-i" and
config="//.clang-format" (to bring the file into the sandbox).